+++ src/db/dexie/db.ts
@@ -15,6 +15,7 @@ import type {
   Setting,
   Task,
   Routine,
+  WeeklyPlan,
 } from "@/shared/types";
 import {
   DEXIE_DATABASE_NAME,
@@ -26,6 +27,7 @@ import {
   DEXIE_SCHEMA_V6,
   DEXIE_SCHEMA_V7,
   DEXIE_SCHEMA_V8,
+  DEXIE_SCHEMA_V9,
   DEXIE_SCHEMA_VERSION,
   DEXIE_SCHEMA_VERSION_3,
   DEXIE_SCHEMA_VERSION_4,
@@ -33,6 +35,7 @@ import {
   DEXIE_SCHEMA_VERSION_6,
   DEXIE_SCHEMA_VERSION_7,
   DEXIE_SCHEMA_VERSION_8,
+  DEXIE_SCHEMA_VERSION_9,
 } from "./schema";
 
 export class AliosDatabase extends Dexie {
@@ -50,6 +53,7 @@ export class AliosDatabase extends Dexie {
   financeTransactions!: Table<FinanceTransaction, string>;
   financeObligations!: Table<FinanceObligation, string>;
   routines!: Table<Routine, string>;
+  weeklyPlans!: Table<WeeklyPlan, string>;
 
   constructor() {
     super(DEXIE_DATABASE_NAME);
@@ -61,6 +65,7 @@ export class AliosDatabase extends Dexie {
     this.version(DEXIE_SCHEMA_VERSION_6).stores(DEXIE_SCHEMA_V6);
     this.version(DEXIE_SCHEMA_VERSION_7).stores(DEXIE_SCHEMA_V7);
     this.version(DEXIE_SCHEMA_VERSION_8).stores(DEXIE_SCHEMA_V8);
+    this.version(DEXIE_SCHEMA_VERSION_9).stores(DEXIE_SCHEMA_V9);
   }
 }